+++ frontend/src/pages/Home.jsx
@@ -2,17 +2,12 @@ import Hero from "../components/Home/Hero"
 import OfferCard from "../components/Home/OfferCard"
 import Subtitle from "../components/Home/Subtitle"
 import Carosel from "../components/Home/Carosel"
-import { useEffect } from "react";
-import { getBooks } from "../store/bookSlice";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import StatusCode from "../utils/StatusCode";
 
 export default function Home() {
   const { data, featuredData, bestSellerData, status } = useSelector(state => state.books);
-  const dispatch = useDispatch();
-  useEffect(() => {
-    dispatch(getBooks());
-  }, [])
+
 
   if (status === StatusCode.LOADING) {
     return <p>Loading...</p>
